+++ src/lib/forms-schema/element-schema.ts
@@ -87,6 +87,9 @@ import AbnElement, { type as abnElementType } from './elements/AbnElement'
 import FreshdeskDependentFieldElement, {
   type as freshdeskDependentFieldType,
 } from './elements/FreshdeskDependentFieldElement'
+import LiquorLicenceElement, {
+  type as liquorLicenceElementType,
+} from './elements/LiquorLicenceElement'
 // Creating an object here so we get a Typescript error when adding a
 // new element type and forgetting to add to the array of allowed types
 const elementTypesMap: Record<
@@ -129,6 +132,7 @@ const elementTypesMap: Record<
   [bsbElementType]: null,
   [abnElementType]: null,
   [freshdeskDependentFieldType]: null,
+  [liquorLicenceElementType]: null,
 }
 
 const typeCase = (type: FormTypes.FormElementType) => {
@@ -251,5 +255,6 @@ const schema = Joi.object({
   .when(typeCase(freshdeskDependentFieldType), {
     then: FreshdeskDependentFieldElement,
   })
+  .when(typeCase(liquorLicenceElementType), { then: LiquorLicenceElement })
 
 export default schema