+++ src/lib/forms-schema/elements/LiquorLicenceElement.ts
@@ -1,0 +1,30 @@
+import Joi from 'joi'
+import {
+  baseSchemas,
+  name,
+  label,
+  hint,
+  requiredSchemas,
+  readOnly,
+  conditionallyShowSchemas,
+  placeholderValue,
+  lookupSchemas,
+  customCssClasses,
+  hintPosition,
+} from '../property-schemas'
+
+export const type = 'apiNSWLiquorLicence'
+
+export default Joi.object({
+  ...baseSchemas,
+  name,
+  label,
+  hint,
+  hintPosition,
+  ...requiredSchemas,
+  readOnly,
+  ...conditionallyShowSchemas,
+  ...lookupSchemas,
+  placeholderValue,
+  customCssClasses,
+})