+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `nswAPILiquorLicence` form element
 - `uploadEmailAttachment` to `Forms` class
 - `emailAttachmentsEndpoint` to email form workflow events validation